+++ lib/src/widgets/completion/completion.dart
@@ -1,1 +1,2 @@
 export 'completion_builder.dart';
+export 'streamable_completion_builder.dart';